Fix some issue with saved defaults (selection mode and snap to grid)
These two preferences are confusing when saved with non-default
values. For example, snap to grid should be almost always turned off
and turned on only when the user do it explicitly.

The other one is the subtraction selection mode, if we save that as a
default mode, users will find that the selection tool stop working
when the program is re-started.

Both of these issues generated user support over all these years.